% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mutations.R
\name{apply_mutations}
\alias{apply_mutations}
\title{Apply random mutations to a community}
\usage{
apply_mutations(
  traits_comm,
  prob_mutation = default_prob_mutation(),
  mutation_sd = default_mutation_sd()
)
}
\arguments{
\item{traits_comm}{numeric vector of variable length, the trait values of
every individual in the community.}

\item{prob_mutation}{numeric between 0 and 1, the probability that any new
individual is sampled with a mutation.}

\item{mutation_sd}{numeric \verb{>= 0}, the standard deviation of the normal
distrbution from which mutations are drawn.}
}
\description{
The trait value of each individual in the input community is modified, with
probability \code{prob_mutation}, by a mutation sampled in a normal distribution
of mean \code{0} and standard deviation \code{mutation_sd}.
}
\author{
Theo Pannetier
}
