% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_speciation.R
\name{apply_speciation}
\alias{apply_speciation}
\title{Resolve cladogenetic and anagetic speciation in the island community}
\usage{
apply_speciation(
  island_comm,
  mainland_comm,
  trait_dist_sp = default_trait_dist_sp()
)
}
\arguments{
\item{island_comm}{a tibble with one row per individual in the community and three
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains species names (characters)
\item \code{ancestral_species} contains acnestral species names (characters)
\item (optionally, if immigration is enabled) \code{root_species} (characters), the
mainland species that this species' clade originates from.
}}

\item{mainland_comm}{a tibble defining what species exist in the mainland and
their trait distribution. One row per species, three columns:
\itemize{
\item \code{species}, (character) the name of the species
\item \code{mean_z}, (numeric) mean value of the trait for that species
\item \code{z_sd}, (numeric) standard deviation of the trait for that species
}}

\item{trait_dist_sp}{numeric, the minimal trait distance between two
clusters of individuals triggering speciation.}
}
\description{
Scan the island community and if conditions for cladogenetic speciation are
met, splits species accordingly. Then does the same for anagenetic speciation.
}
\details{
\strong{For cladogenesis}:
For each species, check if there is any gap \verb{>= trait_dist_sp} in trait
values, and split the relevant species in two when one is found. The less
numerous half becomes the new species.

\strong{For anagenesis}:
Only concerns non-endemic species, i.e. species that are still the same as
one of the populations in \code{mainland_comm}. Such a species becomes a new
species if the \emph{average} trait value of its members differs from the average
trait value of the mainland population by \code{trait_dist_sp}.
}
\note{
\code{apply_speciation()} can currently only split one species into two. If
multiple gaps emerge in a species at a single time step, only the first one
will be treated. As long as branching does not happen at every generation,
this is unlikely to be an issue, because an unresolved gap will be caught on
the next time step. In phylogenetic terms, this results in polytomies being
resolved as soft polytomies.
Besides, simultaneous branching events in different species are handled
perfectly fine.
}
\author{
Théo Pannetier
}
