% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_ltt.R
\name{average_ltt}
\alias{average_ltt}
\title{Compute the average LTT with 95\% CI for a set of trees given a time sequence}
\usage{
average_ltt(ltt_tbls, t_seq)
}
\arguments{
\item{ltt_tbls}{a list of \code{tibble} that contain the LTT of the trees with columns
\code{time} and \code{N}, e.g. the output of \code{\link[=get_ltt_tbl]{get_ltt_tbl()}}.}

\item{t_seq}{a vector of doubles containing the times at which the average N
should be evaluated.}
}
\value{
a \code{tibble} with columns \code{time}, \code{N}, \code{metric}. Values are the mean
number of species, standard deviation, SEM, and 95\% upper and lower CI.
}
\description{
Compute the average LTT with 95\% CI for a set of trees given a time sequence
}
\author{
Theo Pannetier
}
