% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_newick_str_comsie.R
\name{build_newick_tbl}
\alias{build_newick_tbl}
\title{Convert a species history table to a Newick string}
\usage{
build_newick_tbl(spp_tbl)
}
\arguments{
\item{spp_tbl}{a table with phylogenetic information for each species, the
output of \code{\link[=build_comsie_spp_tbl]{build_comsie_spp_tbl()}}}
}
\description{
Reads a tibble recording species birth, death and phylogenetic relations,
figures out the branch lengths and return a phylogenetic tree in Newick
format. Based on \code{\link[DDD:L2phylo]{DDD::L2phylo()}}.
}
\author{
Théo Pannetier
}
