% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_sp_colour.R
\name{cat_sp_colour}
\alias{cat_sp_colour}
\title{Cat the name of a species, coloured by its hex colour code}
\usage{
cat_sp_colour(sp_name, bg = FALSE)
}
\arguments{
\item{sp_name}{character, a 6-digit hex code for a colour that also serves as
name for the species}

\item{bg}{logical, should the colour apply to the background?}
}
\description{
Cat the name of a species, coloured by its hex colour code
}
