% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-check_ddd_version.R
\name{check_ddd_version}
\alias{check_ddd_version}
\title{Check that the installed version of \code{DDD} allows for the functions
required for \code{comrad}}
\usage{
check_ddd_version()
}
\description{
imperfect, for now errors if the installed version of \code{DDD} is not satisfying.
}
\author{
Theo Pannetier
}
