% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mainland_comm.R
\name{create_mainland_comm}
\alias{create_mainland_comm}
\title{Create the mainland community}
\usage{
create_mainland_comm(mainland_nb_species, z_range, mainland_z_sd = 0.2)
}
\arguments{
\item{mainland_nb_species}{integer, the number of species in the island
community.}

\item{z_range}{a range of trait values to sample the mean trait values of
each species from.}

\item{mainland_z_sd}{standard deviation of the trait distribution of mainland
species}
}
\description{
Create a static pool of mainland species to draw immigrants from, with
the parameters of the associated trait distribution
}
\details{
trait values of individuals will be sampled in a normal distribution,
with the mean of each species sampled in a uniform distribution bounded by
\code{z_range}.
}
