% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_loglik_func.R
\name{dd_loglik_func}
\alias{dd_loglik_func}
\title{Log-likelihood function of a diversity-dependent model given waiting times}
\usage{
dd_loglik_func(params, waiting_times_tbl, speciation_func, extinction_func)
}
\arguments{
\item{params}{a named vector containing the values of the parameters of
\code{speciation_func} and \code{extinction_func}.}

\item{waiting_times_tbl}{the output of \code{\link[=waiting_times]{waiting_times()}}, a table containing the
waiting times to a next event, type of events, and species diversity.}

\item{speciation_func}{a function describing how the rate of speciation varies
with the number of species \code{N}. Must take arguments \code{params} and \code{N}. Inside
the function, parameters must be referred as \code{params$param_name}.}

\item{extinction_func}{a function describing how the rate of extinction varies
with the number of species \code{N}. Must take arguments \code{params} and \code{N}. Inside
the function, parameters must be referred as \code{params$param_name}.}
}
\value{
a unique log-likelihood value
}
\description{
Compute the log-likelihood of athe diversity-dependent model specified by
\code{speciation_func}, \code{extinction_func}, and \code{params} given a set of waiting
times to next events and associated diversity.
Meant to be called internally by \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\author{
Theo Pannetier
}
