% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_models.R
\name{dd_model_comrad_to_ddd}
\alias{dd_model_comrad_to_ddd}
\title{Values of \code{DDD} argument \code{ddmodel} corresponding to \code{comrad} DD models}
\usage{
dd_model_comrad_to_ddd(dd_model_name)
}
\arguments{
\item{dd_model_name}{character, the name of the DD model in comrad}
}
\value{
an integer code, value for this DD model in \code{DDD}
}
\description{
Values of \code{DDD} argument \code{ddmodel} corresponding to \code{comrad} DD models
}
\author{
Theo Pannetier
}
