% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_lc.R
\name{dd_model_lc}
\alias{dd_model_lc}
\title{Diversity-dependent model with linear (l) dependence on speciation and
constant (c) extinction}
\usage{
dd_model_lc()
}
\description{
A list specifying a DD model with linear diversity-dependence on the
speciation rate and constant-rate extinction; to be fed as argument
\code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0} - (\lambda_{0} - \mu_{0}) \frac{N}{K} \\ \mu(N) = \mu_{0}}
}
\author{
Theo Pannetier
}
