% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_lp.R
\name{dd_model_lp}
\alias{dd_model_lp}
\title{Diversity-dependent model with linear (l) dependence on speciation and
power (p) dependence on extinction}
\usage{
dd_model_lp()
}
\description{
A list specifying a DD model with linear diversity-dependence on the
speciation rate and power diversity-dependence on the extinction rate;
to be fed as argument \code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
This corresponds to the "exponential" model found in \code{DDD} (\code{ddmodel = 6}) and elsewhere (\code{BAMM}).
}
\details{
\deqn{\lambda(N) = \lambda_{0} - (1 - \alpha)(\lambda_{0} - \mu_{0}) \frac{N}{K}}
\deqn{\mu(N) = \mu_{0}((1 - \alpha) + \alpha \frac{\lambda_{0}}{\mu_{0}})^{\frac{N}{K}}}
}
\author{
Theo Pannetier
}
