% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_pc.R
\name{dd_model_pc}
\alias{dd_model_pc}
\title{Diversity-dependent model with power (p) dependence on speciation and
constant (c) extinction}
\usage{
dd_model_pc()
}
\description{
A list specifying a DD model with power diversity-dependence on the
speciation rate and constant-rate extinction; to be fed as argument
\code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}. This corresponds to the "exponential" model found in \code{DDD} (\code{ddmodel = 2})
}
\details{
\deqn{\lambda(N) = \lambda_{0} \times N^{-\frac{log(\frac{\lambda_{0}}{\mu_{0}})}{log(K)}} \\ \mu(N) = \mu_{0}}
}
\author{
Theo Pannetier
}
