% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_pl.R
\name{dd_model_pl}
\alias{dd_model_pl}
\title{Diversity-dependent model with power (p) dependence on speciation and
linear (l) dependence on extinction}
\usage{
dd_model_pl()
}
\description{
A list specifying a DD model with power diversity-dependence on the
speciation rate and linear diversity-dependence extinction rate;
to be fed as argument \code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
This corresponds to the "exponential" model found in \code{DDD} (\code{ddmodel = 8})
}
\details{
\deqn{\lambda(N) = \lambda_{0}(\alpha + (1 - \alpha) \frac{\mu_{0}}{\lambda_{0}})^{\frac{N}{K}}}
\deqn{\mu(N) = \mu_{0} + \alpha(\lambda_{0} - \mu_{0}) \frac{N}{K}}
}
\author{
Theo Pannetier
}
