% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_px.R
\name{dd_model_px}
\alias{dd_model_px}
\title{Diversity-dependent model with power (p) dependence on speciation and
exponential (x) dependence on extinction}
\usage{
dd_model_px()
}
\description{
A list specifying a DD model with power diversity-dependence on both the
speciation rate and extinction rate; to be fed as argument \code{dd_model} to
\code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}. This corresponds to the "exponential"
model found in \code{DDD} (\code{ddmodel = 15}) and elsewhere (e.g. \code{BAMM}).
}
\details{
\deqn{\lambda(N) = \lambda_{0} \times N^{-\frac{log\Big(\frac{\lambda_0}{\alpha(\lambda_0 - \mu_0) + \mu_0}\Big)}{log(K)}}}
\deqn{\mu(N) = \mu_{0}((1 - \alpha) + \alpha \frac{\lambda_{0}}{\mu_{0}})^{\frac{N}{K}}}
}
\author{
Theo Pannetier
}
