% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_models.R
\name{dd_model_to_extinction_func}
\alias{dd_model_to_extinction_func}
\title{Find DD extinction function associated with a DD model}
\usage{
dd_model_to_extinction_func(dd_model)
}
\arguments{
\item{dd_model}{character, one of \code{\link[=dd_model_names]{dd_model_names()}}}
}
\value{
a character, "constant", "linear", "expo" or "power"
}
\description{
Find DD extinction function associated with a DD model
}
