% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_xc.R
\name{dd_model_xc}
\alias{dd_model_xc}
\title{Diversity-dependent model with exponential (x) dependence on speciation and
constant (c) extinction}
\usage{
dd_model_xc()
}
\description{
A list specifying a DD model with exponential diversity-dependence on the
speciation rate and constant-rate extinction; to be fed as argument
\code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0}(\frac{\mu_{0}}{\lambda_{0}})^{\frac{N}{K}} \\ \mu(N) = \mu_{0}}
}
\author{
Theo Pannetier
}
