% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_xp.R
\name{dd_model_xp}
\alias{dd_model_xp}
\title{Diversity-dependent model with exponential (x) dependence on speciation and
power (p) dependence on extinction}
\usage{
dd_model_xp()
}
\description{
A list specifying a DD model with exponential diversity-dependence on the
speciation rate and power diveristy-dependence on the extinction rate; to be
fed as argument \code{dd_model} to \code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0}(\alpha + (1 - \alpha) \frac{\mu_{0}}{\lambda_{0}})^{\frac{N}{K}}}
\deqn{\mu(N) = \mu_{0} \times N^{\frac{log\Big(\alpha\frac{\lambda_0 - \mu_0}{\lambda_0}+1\Big)}{log(K)}}}
}
\author{
Theo Pannetier
}
