% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_model_xx.R
\name{dd_model_xx}
\alias{dd_model_xx}
\title{Diversity-dependent model with exponential (x) dependence on speciation and
exponential (x) dependence on extinction}
\usage{
dd_model_xx()
}
\description{
A list specifying a DD model with exponential diversity-dependence on both the
speciation rate and extinction rate; to be fed as argument \code{dd_model} to
\code{\link[=fit_dd_model_with_fossil]{fit_dd_model_with_fossil()}}.
}
\details{
\deqn{\lambda(N) = \lambda_{0}(\alpha + (1 - \alpha) \frac{\mu_{0}}{\lambda_{0}})^{\frac{N}{K}}}
\deqn{\mu(N) = \mu_{0}((1 - \alpha) + \alpha \frac{\lambda_{0}}{\mu_{0}})^{\frac{N}{K}}}
}
\author{
Theo Pannetier
}
