% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddd_imports.R
\name{ddd_optimizer}
\alias{ddd_optimizer}
\title{Optimizer functions from package DDD}
\usage{
ddd_optimizer(
  optimpars = c(1e-04, 1e-04, 1e-06, 1000),
  num_cycles = 1,
  fun,
  trparsopt,
  ...
)
}
\arguments{
\item{optimpars}{Parameters of the optimization: relative tolerance in
function arguments, relative tolerance in function value, absolute tolerance
in function arguments, and maximum number of iterations}

\item{num_cycles}{Number of cycles of the optimization. When set to Inf, the
optimization will be repeated until the result is, within the tolerance,
equal to the starting values, with a maximum of 10 cycles.}

\item{fun}{Function to be optimized}

\item{trparsopt}{Initial guess of the parameters to be optimized}

\item{...}{Any other arguments of the function to be optimimzed, or settings
of the optimization routine}
}
\value{
\item{out}{ A list containing optimal function arguments
(\code{par}, the optimal function value (\code{fvalues}) and whether the
optimization converged (\code{conv})}.
}
\description{
See \code{\link[DDD:optimizer]{DDD::optimizer()}} documentation. I have frozen argument \code{optimmethod}
to \code{subplex}, and added an early exit if the first subplex cycle evaluates
to \code{-Inf}.
}
