% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for main parameters of the model}
\usage{
default_params_doc(
  comm,
  trait_ind,
  traits_comm,
  island_comm,
  mainland_comm,
  mainland_nb_species,
  mainland_z_sd,
  competition_sd,
  trait_opt,
  carrying_cap_opt,
  carrying_cap_sd,
  growth_rate,
  prob_mutation,
  mutation_sd,
  trait_dist_sp,
  nb_gens,
  fitness_func,
  comrad_tbl
)
}
\arguments{
\item{comm}{a tibble with one row per individual in the community and three
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains species names (characters)
\item \code{ancestral_species} contains ancestral species names (characters)
\item (optionally, if immigration is enabled) \code{root_species} (characters), the
mainland species that this species' clade originates from.
}}

\item{trait_ind}{numeric. Trait value \code{z} of a focal individual.}

\item{traits_comm}{numeric vector of variable length, the trait values of
every individual in the community.}

\item{island_comm}{a tibble with one row per individual in the community and three
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains species names (characters)
\item \code{ancestral_species} contains acnestral species names (characters)
\item (optionally, if immigration is enabled) \code{root_species} (characters), the
mainland species that this species' clade originates from.
}}

\item{mainland_comm}{a tibble defining what species exist in the mainland and
their trait distribution. One row per species, three columns:
\itemize{
\item \code{species}, (character) the name of the species
\item \code{mean_z}, (numeric) mean value of the trait for that species
\item \code{z_sd}, (numeric) standard deviation of the trait for that species
}}

\item{mainland_nb_species}{integer, the number of species in the island
community.}

\item{mainland_z_sd}{standard deviation of the trait distribution of mainland
species}

\item{competition_sd}{numeric \verb{>= 0}.Parameter \eqn{\sigma_{\alpha}} of
the competition coefficient. Modulates the the strength of competition
between two individuals given their distance in trait space.}

\item{trait_opt}{numeric. The optimal trait value, \eqn{z_{opt}}.
\code{get_carrying_cap(trait_opt) = carrying_cap_opt}.}

\item{carrying_cap_opt}{numeric, value of the carrying capacity at
\code{trait_opt}}

\item{carrying_cap_sd}{numeric \verb{>= 0}. Parameter \eqn{\sigma_K} of the
carrying capacity. Modulates how fast the carrying capacity decays when
moving away from the optimal trait value.}

\item{growth_rate}{numeric \verb{>= 0}, the baseline growth rate. Generations
being discrete, high values will cause chaos.}

\item{prob_mutation}{numeric between 0 and 1, the probability that any new
individual is sampled with a mutation.}

\item{mutation_sd}{numeric \verb{>= 0}, the standard deviation of the normal
distrbution from which mutations are drawn.}

\item{trait_dist_sp}{numeric, the minimal trait distance between two
clusters of individuals triggering speciation.}

\item{nb_gens}{integer, how many generations should the simulation be run
for?}

\item{fitness_func, }{the name of the function used to compute fitness.}

\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}} or read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\author{
Theo Pannetier, based on skeleton stolen from Richel J.C. Bilderbeek.
}
