% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_comsie_pars.R
\name{default_pars}
\alias{default_pars}
\alias{default_mainland_nb_species}
\alias{default_mainland_z_sd}
\title{Default parameter values}
\usage{
default_mainland_nb_species()

default_mainland_z_sd()
}
\description{
Default parameter values
}
