% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_ephemeral_spp.R
\name{exclude_ephemeral_spp}
\alias{exclude_ephemeral_spp}
\title{Exclude ephemeral species}
\usage{
exclude_ephemeral_spp(comrad_tbl, min_nb_gens = 2)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{min_nb_gens}{integer. Species that occur in less sampled generations
than this will be excluded.}
}
\description{
Exclude species sampled for less than a specified number of generations.
}
\details{
Mind that generations not sampled in the table can of course not be
accounted for.
}
\author{
Théo Pannetier
}
