% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dd_adequacy_castor.R
\name{extract_phylo_data}
\alias{extract_phylo_data}
\title{Extract data from phylogeny}
\usage{
extract_phylo_data(phylo, data_type)
}
\arguments{
\item{phylo}{a \code{phylo} object}

\item{data_type}{either \code{"node_age"} or \code{"branch_length"}.}
}
\value{
a vector of the data extracted from the tree.
}
\description{
Helper function to \code{test_dd_adequacy_castor()}. Extracts node ages or edge
lengths from a phylogeny
}
