% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_functions.R
\name{fitness_func}
\alias{fitness_func}
\alias{fitness_func_logistic}
\alias{fitness_func_pontarp}
\alias{fitness_func_ricker}
\title{Fitness functions}
\usage{
fitness_func_logistic(growth_rate, n_eff, carrying_cap)

fitness_func_pontarp(growth_rate, n_eff, carrying_cap)

fitness_func_ricker(growth_rate, n_eff, carrying_cap)
}
\arguments{
\item{growth_rate}{numeric \verb{>= 0}, the baseline growth rate. Generations
being discrete, high values will cause chaos.}

\item{n_eff}{Effective population size experienced by an individual, see
\code{\link[=get_n_eff]{get_n_eff()}}}

\item{carrying_cap}{The carrying capacity experienced by an individual, see
\code{\link[=get_carrying_cap]{get_carrying_cap()}}.}
}
\description{
A set of equations to compute the fitness of an individual. All three
functions specify a decline of fitness with local niche saturation, but
the starting and saturation fitness differ (see Details).
}
\details{
The table below reports the fitness values associated with hallmark levels
of saturation. Note that all three functions are \code{pmax()}'d.
\tabular{cccccc}{
\strong{Neff / K} \tab \strong{0} \tab \strong{1/2} \tab
\strong{1} \tab \strong{2} \tab \strong{infty} \cr
logistic \tab r \tab r/2 \tab 0 \tab 0 \tab 0 \cr
pontarp \tab 1 + r \tab 1 + r/2 \tab 1 \tab 1 - r \tab 0 \cr
ricker \tab exp(r) \tab exp(r/2) \tab 1 \tab exp(-r) \tab 0 \cr
}
}
\author{
Théo Pannetier
}
