% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg_tau.R
\name{get_avg_tau}
\alias{get_avg_tau}
\title{Estimate the diversity-dependent trait evolution rate}
\usage{
get_avg_tau(comrad_tbl)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}} or read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.}
}
\description{
Estimate the average trait evolution rate over each state of
diversity for a \code{comrad} simulation.
}
\author{
Theo Pannetier
}
