% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comp_coeff_comm.R
\name{get_comp_coeff_comm}
\alias{get_comp_coeff_comm}
\title{Get the competition coefficients between an individual and the rest of the
community}
\usage{
get_comp_coeff_comm(
  trait_ind,
  traits_comm,
  competition_sd = default_competition_sd()
)
}
\arguments{
\item{trait_ind}{numeric. Trait value \code{z} of a focal individual.}

\item{traits_comm}{numeric vector of variable length, the trait values of
every individual in the community.}

\item{competition_sd}{numeric \verb{>= 0}.Parameter \eqn{\sigma_{\alpha}} of
the competition coefficient. Modulates the the strength of competition
between two individuals given their distance in trait space.}
}
\value{
a numeric vector containing the competitive effect caused by each
individual in the community on the focal individual. Each effect varies
between 0 (no effect at all) and 1 (the two individuals have the same trait
value).
}
\description{
Returns the competition coefficients experienced by a focal individual from
each individual in the community, including the focal individual on itself.
}
\author{
Theo Pannetier
}
