% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dnltt.R
\name{get_dnltt}
\alias{get_dnltt}
\title{Compute the delta nLTT between two trees}
\usage{
get_dnltt(ltt_1, ltt_2)
}
\arguments{
\item{ltt_1}{a \code{tibble} that contains the LTT of a tree, with columns
\code{time} and \code{N}, e.g. the output of \code{\link[=get_ltt_tbl]{get_ltt_tbl()}}.}

\item{ltt_2}{a \code{tibble} that contains the LTT of another tree, with columns
\code{time} and \code{N}, e.g. the output of \code{\link[=get_ltt_tbl]{get_ltt_tbl()}}.}
}
\value{
a \code{double}, the delta nLTT between the two trees.
}
\description{
Shortcut function calling \code{\link[nLTT:nltt_diff_exact_extinct]{nLTT::nltt_diff_exact_extinct()}}
}
