% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edge_tbl.R
\name{get_edge_tbl}
\alias{get_edge_tbl}
\title{Get a table with node and edge info}
\usage{
get_edge_tbl(phylo)
}
\arguments{
\item{phylo}{an object of the class \code{phylo} as introduced in
\link[ape:read.tree]{ape}. The phylogeny must start with the crown node
(not stem), and be binary (no hard polytomies).}
}
\value{
a \code{tibble} with one observation per branch and five variables:
\itemize{
\item \code{parent_node} ID of the parent node
\item \code{child_node} ID of the child node
\item \code{is_tip} logical, is the (child) node a tip (\code{TRUE}) of the tree or an
internal node (\code{FALSE})
\item \code{edge_length} numeric length of the edge between parent and child,
i.e. the branch length
\item \code{time} for timetrees, the time of the node relative to present (taken as
the latest tip(s) in the tree)
}
}
\description{
Returns a table containing the tree topology (parent-to-child node pairs;
whether a node is a tip or internal), edge (branch) lengths and timing of
nodes (branching times).
}
\author{
Théo Pannetier
}
