% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hat_k.R
\name{get_hat_k}
\alias{get_hat_k}
\title{Estimate equilibrium diversity from LTT table}
\usage{
get_hat_k(avg_ltt_tbl)
}
\arguments{
\item{avg_ltt_tbl}{a \code{tibble} with the average number of species across
replicate simulations, e.g. the output of \code{\link[=avg_ltt]{avg_ltt()}}}
}
\description{
Estimate \code{hat_k}, an empircial measure of equilibrium diversity
at the end of the simulation.
\code{hat_k} is simply computed as the average number of species \code{N} over
replicate simulations and over the last sixth of the simulation time.
}
