% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ltt_tbl.R
\name{get_ltt_tbl}
\alias{get_ltt_tbl}
\title{Get a lineage-through-time table}
\usage{
get_ltt_tbl(phylo)
}
\arguments{
\item{phylo}{an object of the class \code{phylo} as introduced in
\link[ape:read.tree]{ape}. The phylogeny must start with the crown node
(not stem), and be binary (no hard polytomies).
#' @return a \code{tibble} with one observation per node and two variables:
\itemize{
\item \code{time} the time relative to present (taken as the latest tip(s) in
the tree)
\item \code{N} number of lineages or species in the phylogeny at this point in time
}}
}
\description{
Pairs the time of every node in the tree with the number of species
in the tree at that point in time, i.e. lineages-through-time if the tree is
reconstructed, or species-through-time if the tree includes lineages extinct
before present.
}
\author{
Théo Pannetier
}
