% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_eff.R
\name{get_n_eff}
\alias{get_n_eff}
\title{Compute the effective population size}
\usage{
get_n_eff(z, traits_comm, competition_sd = default_competition_sd())
}
\arguments{
\item{z}{numeric, the trait value(s) to get the effective size for. \code{z} and
\code{traits_comm} will usually be the same vector, including during the
simulation. However \code{z} can be different for \code{traits_comm}, if one wants to
know the fitness on the trait axis at points not occupied by any individual,
typically when constructing the fitness landscape.}

\item{traits_comm}{numeric vector, the trait values of all individuals in the
community.}

\item{competition_sd}{numeric \verb{>= 0}.Parameter \eqn{\sigma_{\alpha}} of
the competition coefficient. Modulates the the strength of competition
between two individuals given their distance in trait space.}
}
\description{
Computes \code{n_eff}, the effective population size experienced by an
individual.
}
\details{
\code{n_eff} sums the competitive effects an individual receives from
every individual in the community, including the individual itself. It is
called effective population size because it is the size of the population
that is relevant for competition.
}
\author{
Theo Pannetier
}
