% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_root_age.R
\name{get_root_age}
\alias{get_root_age}
\title{Get the age of the root of a tree}
\usage{
get_root_age(phylo, present_time = 0)
}
\arguments{
\item{phylo}{a \code{phylo} class object}

\item{present_time}{the time at present, assumed to be 0 by default}
}
\value{
the root age, a double.
}
\description{
A bit hacky; not the most efficient function but that's what I could come up with quickly.
}
\author{
Theo Pannetier
}
