% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_viable_z_range.R
\name{get_viable_z_range}
\alias{get_viable_z_range}
\title{Given K(z), return the min and max values of z viable for a population of some size}
\usage{
get_viable_z_range(pop_size, trait_opt, carrying_cap_opt, carrying_cap_sd)
}
\arguments{
\item{pop_size}{integer, a population size}

\item{trait_opt}{numeric. The optimal trait value, \eqn{z_{opt}}.
\code{get_carrying_cap(trait_opt) = carrying_cap_opt}.}

\item{carrying_cap_opt}{numeric, value of the carrying capacity at
\code{trait_opt}}

\item{carrying_cap_sd}{numeric \verb{>= 0}. Parameter \eqn{\sigma_K} of the
carrying capacity. Modulates how fast the carrying capacity decays when
moving away from the optimal trait value.}
}
\value{
double, the max value of z that can support a population of size \code{pop_size}
}
\description{
Given K(z), return the min and max values of z viable for a population of some size
}
