% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_brute_force_opt.R
\name{has_brute_force_opt}
\alias{has_brute_force_opt}
\title{Are brute force options available?}
\usage{
has_brute_force_opt()
}
\description{
\code{comrad} has some "brute force" options (namely, using OpenMP and/or SIMD) to
speed up the simulations, but these break the build because of portability
issues. To preserve the build on \code{master}, these options are only enabled on
separate branch \code{brute_force}. This function returns whether brute force is
enabled on branch currently in use.
}
\author{
Théo Pannetier
}
