% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_ltt.R
\name{interpolate_ltt}
\alias{interpolate_ltt}
\title{Interpolate LTT for a given time sequence}
\usage{
interpolate_ltt(ltt_tbl, t_seq)
}
\arguments{
\item{ltt_tbl}{a \code{tibble} with columns \code{time} and \code{N}, e.g. the output of
\code{\link[=get_ltt_tbl]{get_ltt_tbl()}}.}

\item{t_seq}{numeric vector, a sequence of time points.}
}
\value{
a \code{tibble} with columns \code{time} and \code{N}, where \code{time} corresponds to every time in \code{t_seq}
}
\description{
Given a lineage-through-time table and a time sequence, this function evaluates the value
of N at every time point in the sequence.
}
\author{
Theo Pannetier
}
