% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt_tbl_to_func.R
\name{ltt_tbl_to_func}
\alias{ltt_tbl_to_func}
\title{Transform the LTT table into a step function}
\usage{
ltt_tbl_to_func(ltt_tbl)
}
\arguments{
\item{ltt_tbl}{a \code{tibble} with columns \code{time} and \code{N}, e.g. the output of
\code{\link[=get_ltt_tbl]{get_ltt_tbl()}}.}
}
\value{
a stepwise function of time that returns the number of lineages
in the tree at any point in time
}
\description{
Converts the lineage-through-time table of a tree into a stepwise function of
time with \code{\link[stats:stepfun]{stats::stepfun()}}
}
\author{
Théo Pannetier
}
