% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aicw_pie.R
\name{plot_aicw_pie}
\alias{plot_aicw_pie}
\title{Plot Akaike weights as a pie chart}
\usage{
plot_aicw_pie(aicw_tbl, slice_by = "dd_model", by_tree = FALSE)
}
\arguments{
\item{aicw_tbl}{a data frame}

\item{slice_by}{either "dd_model" or "speciation_func", "extinction_func",
how the AICw scores should be summarised and the pie chart coloured.}

\item{by_tree}{logical, does the data contain results for each tree separately (\code{TRUE})
or for a set of replicate trees as a whole (\code{FALSE})?}
}
\description{
Plot Akaike weights as a pie chart
}
