% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_bubbles.R
\name{plot_comm_bubbles}
\alias{plot_comm_bubbles}
\title{Plot population densities and trait distribution over time, with bubbles}
\usage{
plot_comm_bubbles(
  comrad_tbl,
  generation_range = c(0, Inf),
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{generation_range}{numeric vector with length 2, supplying the first and
last generation to plot from the dataset. Defaults to every generation in the
table}

\item{xlim}{numeric vector, the range of x-values, passed to
\code{ggplot::coord_cartesian()} if not \code{NULL}.}

\item{ylim}{numeric vector, the range of y-values, passed to
\code{ggplot::coord_cartesian()} if not \code{NULL}.}
}
\description{
Produces a plot of trait evolution where individuals are binned as bubbles
which colour indicate the species identity and the size the number of
individuals in the bin. Plots both species identity and individual densities,
but at the expense of resolution (compared to \code{\link[=plot_comm_trait_evolution]{plot_comm_trait_evolution()}}).
}
\author{
Théo Pannetier
}
