% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_size.R
\name{plot_comm_size}
\alias{plot_comm_size}
\title{Plot community size over generations}
\usage{
plot_comm_size(comrad_tbl, colouring = "species_names", which_geom = "area")
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{colouring}{character specifying how to colour each species. Default
\code{auto} lets \code{ggplot} pick colours as usual. \code{species_names} is used when
colour hexes are sampled as species names during the simulation, and will
colour each species with its colour. This ensures consistency across plots
but may make colours harder to distinguish based on which hexes where
sampled.}

\item{which_geom}{character specifying the geom to plot. \code{"area"} plots a
stack of area plots coloured by species,  \code{"line"} plots the population sizes
of each species in separate lines (still coloured by species).}
}
\description{
Plots the number of individuals in the community over time.
}
\author{
Théo Pannetier
}
