% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_trait_evolution.R
\name{plot_comm_trait_evolution}
\alias{plot_comm_trait_evolution}
\title{Plot the evolution of trait values in the community over generations}
\usage{
plot_comm_trait_evolution(
  comrad_tbl,
  generation_range = c(0, Inf),
  xgrain = 200,
  ygrain = 0.01,
  xlim = NULL,
  ylim = NULL,
  alpha = 0.95,
  hex_fill = "species"
)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{generation_range}{numeric vector with length 2, supplying the first and
last generation to plot from the dataset. Defaults to every generation in the
table}

\item{xgrain}{numeric, the width of  hexes on the x-axis (generations)}

\item{ygrain}{numeric, the width of  hexes on the y-axis (trait)}

\item{xlim}{numeric vector, the range of x-values, passed to
\code{ggplot::coord_cartesian()} if not \code{NULL}.}

\item{ylim}{numeric vector, the range of y-values, passed to
\code{ggplot::coord_cartesian()} if not \code{NULL}.}

\item{hex_fill}{character with two options, \code{"counts"} and \code{"species"}.
\code{"counts"} colours hexes by the count of individuals, and \code{"species"} by
species identity.}
}
\description{
Produces a hex-density plot of all trait values in the community over
generations.
}
\author{
Théo Pannetier
}
