% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_traits.R
\name{plot_comm_traits}
\alias{plot_comm_traits}
\title{Plot trait distribution in a generation}
\usage{
plot_comm_traits(comrad_tbl, generation, binwidth = 0.01)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{generation}{numeric, the index of the generation to plot}

\item{binwidth}{numeric, the width used for binning.}
}
\description{
Plots a histogram of the distribution of trait values in a single generation.
}
\author{
Théo Pannetier
}
