% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comm_traits_anim.R
\name{plot_comm_traits_anim}
\alias{plot_comm_traits_anim}
\title{Plot an animation of trait distribution through generations}
\usage{
plot_comm_traits_anim(
  comrad_tbl,
  generation_range = c(0, Inf),
  binwidth = 0.01
)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}

\item{generation_range}{numeric vector with length 2, supplying the first and
last generation to plot from the dataset. Defaults to every generation in the
table}

\item{binwidth}{numeric, the width used for binning.}
}
\description{
Animated version of \code{\link[=plot_comm_traits]{plot_comm_traits()}}, lopping through generations. Dope.
}
\author{
Théo Pannetier
}
