% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_dd_functions.R
\name{plot_compare_dd_functions}
\alias{plot_compare_dd_functions}
\title{Quickly compare the three diversity-dependent functions}
\usage{
plot_compare_dd_functions(params, N)
}
\arguments{
\item{params}{a numeric vector with four named elements, \code{"alpha_0"}, \verb{"mu_0},
\code{"k"} and \code{"alpha"}.}

\item{N}{a numeric vector, range of values of N to plot}
}
\description{
Draws a quick plot of linear, exponential and power speciation and extinction
functions for the requested parameter values and range of N
}
