% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_models.R
\name{plot_dd_model_colours}
\alias{plot_dd_model_colours}
\title{Plot a key to the colours used to represent DD models}
\usage{
plot_dd_model_colours()
}
\description{
Plot a key to the colours used to represent DD models
}
