% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitness_landscape.R
\name{plot_fitness_landscape}
\alias{plot_fitness_landscape}
\title{Plot the fitness landscape at a given time in the simulation}
\usage{
plot_fitness_landscape(
  comrad_tbl,
  generation,
  z_seq = NULL,
  competition_sd,
  carrying_cap_sd
)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}} or read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.}

\item{generation}{numeric, a single generation step in the simulation.}

\item{z_seq}{numeric vector, the sequence of trait values along which the
fitness landscape should be computed.
Default \code{z_seq = NULL} sets the sequence to every \code{0.01} between twice the
minimal and maximal value in \code{comrad_tbl}}

\item{competition_sd}{numeric \verb{>= 0}.Parameter \eqn{\sigma_{\alpha}} of
the competition coefficient. Modulates the the strength of competition
between two individuals given their distance in trait space.}

\item{carrying_cap_sd}{numeric \verb{>= 0}. Parameter \eqn{\sigma_K} of the
carrying capacity. Modulates how fast the carrying capacity decays when
moving away from the optimal trait value.}
}
\description{
Computes the fitness landscape along trait range \code{z_seq} at time \code{generation}
in the simulation
}
\author{
Théo Pannetier
}
