% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitness_landscape_evolution.R
\name{plot_fitness_landscape_evol}
\alias{plot_fitness_landscape_evol}
\title{Plot the fitness landscape in a \code{comrad} simulation over time}
\usage{
plot_fitness_landscape_evol(
  comrad_tbl,
  z_seq = NULL,
  generation_range = c(0, Inf),
  fitness_lim = c(0, get_fitness(0)),
  competition_sd,
  carrying_cap_sd
)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}} or read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.}

\item{z_seq}{numeric vector, the sequence of trait values along which the
fitness landscape should be computed.
Default \code{z_seq = NULL} sets the sequence to every \code{0.01} between twice the
minimal and maximal value in \code{comrad_tbl}.}

\item{generation_range}{numeric vector with length 2, supplying the first and
last generation to plot from the dataset. Defaults to every generation in the
table}

\item{fitness_lim}{numeric vector of length 2, the plotting range for the
z-axis, the fitness.}

\item{competition_sd}{numeric \verb{>= 0}.Parameter \eqn{\sigma_{\alpha}} of
the competition coefficient. Modulates the the strength of competition
between two individuals given their distance in trait space.}

\item{carrying_cap_sd}{numeric \verb{>= 0}. Parameter \eqn{\sigma_K} of the
carrying capacity. Modulates how fast the carrying capacity decays when
moving away from the optimal trait value.}
}
\description{
Computes the fitness landscape along the time sequence defined by
\code{generation_range} in a 3D plot.
}
\author{
Théo Pannetier
}
