% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_disparity.R
\name{plot_trait_disparity}
\alias{plot_trait_disparity}
\title{Plot trait disparity over time}
\usage{
plot_trait_disparity(comrad_tbl, generation_range = c(0, Inf))
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}} or read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.}

\item{generation_range}{numeric vector with length 2, supplying the first and
last generation to plot from the dataset. Defaults to every generation in the
table}
}
\description{
Compute and plot the variance in trait values in the community at every time
step
}
\author{
Théo Pannetier
}
