% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zt.R
\name{plot_zt}
\alias{plot_zt}
\title{Plot trait distribution through time}
\usage{
plot_zt(
  comsie_tbl,
  fill_by = "species",
  xlim,
  ylim,
  binwidths = c(100, 0.01),
  alpha = 1
)
}
\arguments{
\item{comsie_tbl}{a tibble, e.g. output of \code{\link[=read_comsie_tbl]{read_comsie_tbl()}}}

\item{fill_by}{How should rectangles be coloured? either "species" or "clade".}

\item{xlim}{length 2 num vector to pass to ggplot's \code{coord_cartesian()}}

\item{ylim}{length 2 num vector to pass to ggplot's \code{coord_cartesian()}}

\item{binwidths}{length 2 num vector, horizontal and vertical widths of the
rectangles.}

\item{alpha}{the ggplot parameter}
}
\description{
Updated version of \code{\link[comrad:plot_comm_traits_evolution]{comrad::plot_comm_traits_evolution()}}.
Main difference is the (faster) use of rectangles instead of hexes to
represent occupied areas of the trait distribution, and more straightforwards
plotting options.
}
