% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_comm.R
\name{rand_comm}
\alias{rand_comm}
\title{Sample a random \code{comrad} community}
\usage{
rand_comm()
}
\description{
Sample 100 individuals pooled in up to 5 random species, with
trait values sampled in a normal distribution. For testing purposes.
}
\author{
Théo Pannetier
}
