% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_from_dd_model.R
\name{rates_from_dd_model}
\alias{rates_from_dd_model}
\title{Get speciation and extinction rates values given a DD model and parameters}
\usage{
rates_from_dd_model(N_seq, dd_model, params)
}
\arguments{
\item{N_seq}{numeric vector, the values for which rate values are desired}

\item{dd_model}{a list with five named elements that together specify the
diversity-dependent model:
\itemize{
\item \code{name} a two-letter code, the name of the model. First letter specifies the
form of the speciation function, second letter the form of the extinction
function: "l" for "linear", "x" for exponential, "c" for constant.
\item \code{speciation_func}, a function specifying the diversity-dependent speciation
rate. Must take arguments \code{params} and \code{N}.
\item \code{extinction_func}, a function specifying the diversity-dependent extinction
rate. Must take arguments \code{params} and \code{N}.
\item \code{constraints} a list of conditions that parameter values must satisfy. Each
element is a function that takes arguments \code{params} and \code{...}, and returns
\code{TRUE} if the constraint is satisfied, \code{FALSE} if it isn't.
\item \code{params_check} a function that controls the format of \code{params}. Returns an
error if the elements of \code{params} are named differently from what is expected
or if the length differs from the expectation.
}

\code{comrad} contains several \code{dd_model} functions, see for example
\code{\link[=dd_model_lc]{dd_model_lc()}}.}

\item{params}{a named vector containing the values of the parameters of
\code{speciation_func} and \code{extinction_func}.}
}
\value{
a \code{tibble} with variables \code{N}, \code{rate} (speciation or extinction),
\code{value} (value of that rate), and \code{dd_model} (name of the DD model)
}
\description{
Get speciation and extinction rates values given a DD model and parameters
}
\author{
Théo Pannetier
}
