% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates_from_exp_dist.R
\name{rates_from_exp_dist}
\alias{rates_from_exp_dist}
\title{Estimate diversity-dependent rates from mean parameter of the exponential distribution}
\usage{
rates_from_exp_dist(waiting_times_tbl)
}
\arguments{
\item{waiting_times_tbl}{a table with waiting times, the output of
\code{\link[=waiting_times]{waiting_times()}} for one or more phylogenies.}
}
\value{
a \code{tibble} with a row for every possible number of species and 4
columns:
\itemize{
\item \code{N}, level of diversity, nb of species
\item \code{rate} (speciation or extinction)
\item \code{value} value of that rate
\item \code{dd_model} = \code{none}, for compatibility with \code{\link[=rates_from_dd_model]{rates_from_dd_model()}}
\item \code{ci_upper} upper bound of the confidence interval of \code{value}
\item \code{ci_lower} lower bound of the confidence interval of \code{value}
}


}
\description{
For a list of phylogenies, reconstruct the \emph{average, per-capita} rates of
speciation and extinction corresponding to each level of species diversity.
}
\author{
Théo Pannetier
}
