% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_comrad_tbl.R
\name{read_comrad_tbl}
\alias{read_comrad_tbl}
\title{Read simulation data}
\usage{
read_comrad_tbl(path_to_file, skip = 20)
}
\arguments{
\item{path_to_file}{character, path to the \code{.csv} file.}

\item{skip}{numeric, number of lines (of metadata) to skip. Passed to
\code{\link[readr:read_delim]{readr::read_csv()}}.}
}
\description{
Reads a \code{.csv} file as produced by \code{\link[=run_simulation]{run_simulation()}} and returns the parsed
table it contains.
}
\author{
Théo Pannetier
}
