% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_comsie_tbl.R
\name{read_comsie_tbl}
\alias{read_comsie_tbl}
\title{Read simulation data}
\usage{
read_comsie_tbl(path_to_file, skip = NULL)
}
\arguments{
\item{path_to_file}{character, path to the \code{.csv} file.}

\item{skip}{numeric, number of lines (of metadata) to skip. Passed to
\code{\link[readr:read_delim]{readr::read_csv()}}. If left NULL, the function will use a regexp to find
where the simulation output starts.}
}
\description{
Reads a \code{.csv} file as produced by \code{\link[=run_simulation]{run_simulation()}} and returns the parsed
table it contains.
}
\author{
Théo Pannetier
}
