% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events_from_log.R
\name{read_events_from_log}
\alias{read_events_from_log}
\title{Read the sequence of events in a simulation from its log}
\usage{
read_events_from_log(path_to_log)
}
\arguments{
\item{path_to_log}{character, path to the log file}
}
\description{
Read the sequence of events in a simulation from its log
}
