% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_immigration_from_log.R
\name{read_immigration_from_log}
\alias{read_immigration_from_log}
\title{Read immigration events and their time from a simulation's log}
\usage{
read_immigration_from_log(path_to_log)
}
\arguments{
\item{path_to_log}{character, path to the log of the simulation}
}
\value{
a tibble with columns "t" and "species"
}
\description{
Read immigration events and their time from a simulation's log
}
