% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_to_dd_rates.R
\name{sim_to_dd_rates}
\alias{sim_to_dd_rates}
\title{Estimate speciation, extinction, and trait evo rates as function of diversity}
\usage{
sim_to_dd_rates(comrad_tbls)
}
\arguments{
\item{comrad_tbls}{a list containing the output of a set of simulations,
as read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}}
}
\description{
Computes estimates of the speciation, extinction and trait evolution rates
for each value of species diversity.
}
\author{
Theo Pannetier
}
