% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_to_phylo.R
\name{sim_to_phylo}
\alias{sim_to_phylo}
\title{Convert \code{comrad} simulation results into a phylogeny}
\usage{
sim_to_phylo(comrad_tbl, include_stem = TRUE, with_extinct = TRUE)
}
\arguments{
\item{comrad_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}} or read by \code{\link[=read_comrad_tbl]{read_comrad_tbl()}}.}

\item{include_stem}{logical, should the stem lineage and stem age be
included? If not, the string stops at crown lineages.}

\item{with_extinct}{logical, should extinct taxa be included? \code{TRUE} returns
the full tree, \code{FALSE} returns the reconstructed tree.}
}
\description{
Reads a table of results and returns a \code{phylo} class object.
}
\author{
Théo Pannetier
}
