% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_or_crown.R
\name{stem_or_crown}
\alias{stem_or_crown}
\title{Find wheter origin is stem or crown lineages}
\usage{
stem_or_crown(phylo)
}
\arguments{
\item{phylo}{an object of the class \code{phylo} as introduced in
\link[ape:read.tree]{ape}. The phylogeny must start with the crown node
(not stem), and be binary (no hard polytomies).#'}
}
\description{
\code{phylo} is rooted on its stem if nb nodes = nb tips, and on its crown
otherwise
}
\author{
Théo Pannetier
}
