% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_aicw_over_trees.R
\name{summarise_aicw_over_trees}
\alias{summarise_aicw_over_trees}
\title{Average AICw scores over all replicate trees}
\usage{
summarise_aicw_over_trees(aicw_tbl)
}
\arguments{
\item{aicw_tbl}{a data frame with AICw scores for each model and tree}
}
\description{
Only relevant if DD models have been fit separately on each tree, i.e. the
input \code{aicw_tbl} must contain a variable \code{tree}.
}
