% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_waiting_times.R
\name{summarise_waiting_times}
\alias{summarise_waiting_times}
\title{Summarise waiting times data set}
\usage{
summarise_waiting_times(waiting_times_tbl, summary_func = mean)
}
\arguments{
\item{waiting_times_tbl}{the output of \code{\link[=waiting_times]{waiting_times()}} for several trees,
with an extra column \code{replicate} that indicates which tree each waiting time is from.}

\item{summary_func}{character,  the function to summarise the data with. Either
\code{mean} or \code{median}.}
}
\value{
a \code{tibble} with variables \code{replicate}, \code{next_event},
\code{N}, \code{waiting_time} and \code{is_speciation}
}
\description{
Summarise waiting times observations using a summary function
(either mean or median), to get one single observation per tree, value of N
and type of event (speciation or extinction)
}
\author{
Theo Pannetier
}
