% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dd_adequacy_castor.R
\name{test_dd_adequacy_castor}
\alias{test_dd_adequacy_castor}
\title{Test model adequacy using the tests of Louca & Pennell (2021)}
\usage{
test_dd_adequacy_castor(phylos_empirical, phylos_bootstrap, data_type)
}
\arguments{
\item{phylos_empirical}{a list of empirical (i.e. simulated under comrad) trees.}

\item{phylos_bootstrap}{a list of bootstrap (i.e. simulated under a DD model) trees.}

\item{data_type}{the type of data to consider for the test, either \code{"edge_length"} or \code{"node_age"}.}
}
\value{
a vector of "p-values" for each "empirical" tree. Those metrics are not really p-values but fit
the same role: how likely to observe this CDF assuming the empirical trees were simulated under the same model as
the bootstrap trees?
}
\description{
Test whether a DD model is consistent with a set of "empirical" (i.e., comrad) trees.
The test compares the CDF for either node ages or edge lengths of each empirical tree and the
set of bootstrap trees.
}
\author{
Theo Pannetier
}
