% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dd_adequacy_imbalance.R
\name{test_dd_adequacy_imbalance}
\alias{test_dd_adequacy_imbalance}
\title{Test DD model adequacy on tree imbalance}
\usage{
test_dd_adequacy_imbalance(phylos_empirical, phylos_bootstrap, type = "Sackin")
}
\arguments{
\item{phylos_empirical}{a list of empirical (i.e. simulated under comrad) trees.}

\item{phylos_bootstrap}{a list of bootstrap (i.e. simulated under a DD model) trees.}

\item{type}{the index to use to measure tree imbalance. Either \code{"Sackin"} or \code{"Colless"}.}
}
\value{
a list with four elements:
\itemize{
\item \code{delta_imbalance_empirical} vector with the absolute difference in tree
imbalance between every tree in the empirical set and the average bootstrap LTT
\item \code{delta_imbalance_bootstrap} vector with the absolute difference in tree
imbalance  between every tree in the bootsrap set and the average bootstrap LTT
\item \code{p} a measure akin to a p-value, computed as the proportion of bootstrap
trees for which \code{delta_imbalance_bootstrap} > \code{delta_imbalance_empirical}.
This measure is computed once for every tree in the empirical set.
\item \code{ks} output of a two-sample, one-sided (empirical > bootstrap)
Kolmogorov-Smirnov test on the distribution of the delta imbalance of the
empirical and bootsrap test.
}
}
\description{
Test whether the imbalalance pattern of a set of "empirical" (comrad) trees
is consistent with that predicted by a DD model using the Sackin or Colless indices,
via \code{\link[castor:tree_imbalance]{castor::tree_imbalance()}}.
}
\author{
Theo Pannetier
}
