% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wt_vs_pexp.R
\name{test_wt_vs_pexp}
\alias{test_wt_vs_pexp}
\title{Run a one-sample KS test on a sample of waiting times}
\usage{
test_wt_vs_pexp(waiting_times_tbl)
}
\arguments{
\item{waiting_times_tbl}{a \code{tibble}, the output of \code{\link[=waiting_times]{waiting_times()}} for one
or more trees.}
}
\value{
a \code{tibble} with:
\itemize{
\item \code{N} standing diversity
\item \code{event} either speciation or extinction
\item \code{n} number of observations, i.e. waiting times
\item \code{loglik} log-likelihood of the exponential with ML estimate \code{rate}
\item \code{rate} ML estimate of the exponential dist. rate parameter
\item \code{stat} Kolmogorov-Smirnov distance
\item \code{p} p-value of the test
\item \code{test} type of KS test
}
}
\description{
For each value of N and speciation or exinction events,
fits an exponential distribution (with ML) to the waiting times and run a
Kolmogorov-Smirnov test of the waiting times vs the theoretical exponential distribution.
}
\author{
Theo Pannetier
}
