% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_models.R
\name{update_dd_model_name}
\alias{update_dd_model_name}
\title{Update power DD model names}
\usage{
update_dd_model_name(x)
}
\description{
Power DD functions used to be called "x2" ("alternative" exponential,
in contrast to the true exponential functions referred to by "x). This
function updates the names, useful if dealing with data produced by older
versions of \code{comrad}.
}
\details{
@param x, a \code{dd_model} name

@return an updated \code{dd_model} name

@author Théo Pannetier
@export
}
