% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_newick_str.R
\name{write_newick_str}
\alias{write_newick_str}
\title{Convert a species history table to a Newick string}
\usage{
write_newick_str(spp_tbl, include_stem = TRUE)
}
\arguments{
\item{spp_tbl}{a table with phylogenetic information for each species, the
output of \code{\link[=build_spp_tbl]{build_spp_tbl()}}}

\item{include_stem}{logical, should the stem lineage and stem age be
included? If not, the string stops at crown lineages.}
}
\description{
Reads a tibble recording species birth, death and phylogenetic relations,
figures out the branch lenghts and return a phylogenetic tree in Newick
format. Based on \code{\link[DDD:L2phylo]{DDD::L2phylo()}}.
}
\author{
Théo Pannetier
}
