% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_pars.R
\name{default_pars}
\alias{default_pars}
\alias{default_competition_sd}
\alias{default_trait_opt}
\alias{default_carrying_cap_opt}
\alias{default_carrying_cap_sd}
\alias{default_growth_rate}
\alias{default_prob_mutation}
\alias{default_mutation_sd}
\alias{default_init_comm}
\alias{default_seed}
\alias{default_sampling_frac}
\alias{default_trait_dist_sp}
\title{Default parameter values}
\usage{
default_competition_sd()

default_trait_opt()

default_carrying_cap_opt()

default_carrying_cap_sd()

default_growth_rate()

default_prob_mutation()

default_mutation_sd()

default_init_comm()

default_seed()

default_sampling_frac()

default_trait_dist_sp()
}
\description{
Return default values of the simulation parameters, i.e. the default settings
used in Pontarp et al. (2012).
}
\author{
Theo Pannetier
}
