% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-testargs.R
\name{testargs}
\alias{testargs}
\alias{test_island_comm}
\alias{test_comsie_tbl}
\title{Assert a function argument is of correct type}
\usage{
test_island_comm(comm)

test_comsie_tbl(comsie_tbl)
}
\arguments{
\item{comm}{a tibble with one row per individual in the community and three
columns:
\itemize{
\item \code{z} contains (numeric) traits values
\item \code{species} contains species names (characters)
\item \code{ancestral_species} contains ancestral species names (characters)
\item \code{founder} (characters), the mainland species that this species' clade originates from.
}}

\item{comsie_tbl}{a tibble containing the output of a \code{comrad} simulation,
as produced by \code{\link[=run_simulation]{run_simulation()}}.}
}
\description{
Assert a function argument is of correct type
}
\author{
Theo Pannetier
}
